// src/apispec/src/models/ActivateRunRevisionRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sandbox0 API
 * Public HTTP APIs exposed by Sandbox0 regional gateways and the global gateway.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ActivateRunRevisionRequest
 */
export interface ActivateRunRevisionRequest {
    /**
     * 
     * @type {string}
     * @memberof ActivateRunRevisionRequest
     */
    revisionId: string;
}

/**
 * Check if a given object implements the ActivateRunRevisionRequest interface.
 */
export function instanceOfActivateRunRevisionRequest(value: object): value is ActivateRunRevisionRequest {
    if (!('revisionId' in value) || value['revisionId'] === undefined) return false;
    return true;
}

export function ActivateRunRevisionRequestFromJSON(json: any): ActivateRunRevisionRequest {
    return ActivateRunRevisionRequestFromJSONTyped(json, false);
}

export function ActivateRunRevisionRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ActivateRunRevisionRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'revisionId': json['revision_id'],
    };
}

export function ActivateRunRevisionRequestToJSON(json: any): ActivateRunRevisionRequest {
    return ActivateRunRevisionRequestToJSONTyped(json, false);
}

export function ActivateRunRevisionRequestToJSONTyped(value?: ActivateRunRevisionRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'revision_id': value['revisionId'],
    };
}