// src/apispec/src/models/Run.ts
export interface Run {
    /**
     * 
     * @type {string}
     * @memberof Run
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Run
     */
    teamId: string;
    /**
     * 
     * @type {string}
     * @memberof Run
     */
    createdBy?: string;
    /**
     * 
     * @type {string}
     * @memberof Run
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof Run
     */
    slug: string;
    /**
     * 
     * @type {string}
     * @memberof Run
     */
    domainLabel: string;
    /**
     * 
     * @type {string}
     * @memberof Run
     */
    url?: string;
    /**
     * 
     * @type {string}
     * @memberof Run
     */
    activeRevisionId?: string;
    /**
     * 
     * @type {boolean}
     * @memberof Run
     */
    enabled: boolean;
    /**
     * 
     * @type {RunScalePolicy}
     * @memberof Run
     */
    scale: RunScalePolicy;
    /**
     * 
     * @type {Date}
     * @memberof Run
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof Run
     */
    updatedAt: Date;
}