// src/apispec/src/models/RunDeployRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sandbox0 API
 * Public HTTP APIs exposed by Sandbox0 regional gateways and the global gateway.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { RunSource } from './RunSource';
import {
    RunSourceFromJSON,
    RunSourceFromJSONTyped,
    RunSourceToJSON,
    RunSourceToJSONTyped,
} from './RunSource';
import type { RunScalePolicy } from './RunScalePolicy';
import {
    RunScalePolicyFromJSON,
    RunScalePolicyFromJSONTyped,
    RunScalePolicyToJSON,
    RunScalePolicyToJSONTyped,
} from './RunScalePolicy';
import type { RunRevisionSpec } from './RunRevisionSpec';
import {
    RunRevisionSpecFromJSON,
    RunRevisionSpecFromJSONTyped,
    RunRevisionSpecToJSON,
    RunRevisionSpecToJSONTyped,
} from './RunRevisionSpec';

/**
 * 
 * @export
 * @interface RunDeployRequest
 */
export interface RunDeployRequest {
    /**
     * 
     * @type {string}
     * @memberof RunDeployRequest
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof RunDeployRequest
     */
    slug?: string;
    /**
     * 
     * @type {RunScalePolicy}
     * @memberof RunDeployRequest
     */
    scale?: RunScalePolicy;
    /**
     * 
     * @type {RunSource}
     * @memberof RunDeployRequest
     */
    source?: RunSource;
    /**
     * 
     * @type {RunRevisionSpec}
     * @memberof RunDeployRequest
     */
    spec?: RunRevisionSpec;
    /**
     * 
     * @type {boolean}
     * @memberof RunDeployRequest
     */
    activate?: boolean;
}

/**
 * Check if a given object implements the RunDeployRequest interface.
 */
export function instanceOfRunDeployRequest(value: object): value is RunDeployRequest {
    return true;
}

export function RunDeployRequestFromJSON(json: any): RunDeployRequest {
    return RunDeployRequestFromJSONTyped(json, false);
}

export function RunDeployRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): RunDeployRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'slug': json['slug'] == null ? undefined : json['slug'],
        'scale': json['scale'] == null ? undefined : RunScalePolicyFromJSON(json['scale']),
        'source': json['source'] == null ? undefined : RunSourceFromJSON(json['source']),
        'spec': json['spec'] == null ? undefined : RunRevisionSpecFromJSON(json['spec']),
        'activate': json['activate'] == null ? undefined : json['activate'],
    };
}

export function RunDeployRequestToJSON(json: any): RunDeployRequest {
    return RunDeployRequestToJSONTyped(json, false);
}

export function RunDeployRequestToJSONTyped(value?: RunDeployRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'slug': value['slug'],
        'scale': RunScalePolicyToJSON(value['scale']),
        'source': RunSourceToJSON(value['source']),
        'spec': RunRevisionSpecToJSON(value['spec']),
        'activate': value['activate'],
    };
}