// src/apispec/src/models/RunRevision.ts
export interface RunRevision {
    /**
     * 
     * @type {string}
     * @memberof RunRevision
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof RunRevision
     */
    runId: string;
    /**
     * 
     * @type {string}
     * @memberof RunRevision
     */
    teamId: string;
    /**
     * 
     * @type {number}
     * @memberof RunRevision
     */
    number: number;
    /**
     * 
     * @type {RunSource}
     * @memberof RunRevision
     */
    source: RunSource;
    /**
     * 
     * @type {RunRevisionSpec}
     * @memberof RunRevision
     */
    spec: RunRevisionSpec;
    /**
     * 
     * @type {RunRevisionStatus}
     * @memberof RunRevision
     */
    status: RunRevisionStatus;
    /**
     * 
     * @type {string}
     * @memberof RunRevision
     */
    runtimeSandboxId?: string;
    /**
     * 
     * @type {string}
     * @memberof RunRevision
     */
    runtimeClusterId?: string;
    /**
     * 
     * @type {string}
     * @memberof RunRevision
     */
    runtimeContextId?: string;
    /**
     * 
     * @type {Date}
     * @memberof RunRevision
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof RunRevision
     */
    activatedAt?: Date;
}