// src/apispec/src/models/RunRevisionMount.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sandbox0 API
 * Public HTTP APIs exposed by Sandbox0 regional gateways and the global gateway.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface RunRevisionMount
 */
export interface RunRevisionMount {
    /**
     * 
     * @type {string}
     * @memberof RunRevisionMount
     */
    snapshotId: string;
    /**
     * 
     * @type {string}
     * @memberof RunRevisionMount
     */
    mountPath: string;
    /**
     * Run snapshot mounts are immutable and are materialized as read-only volumes at runtime.
     * @type {boolean}
     * @memberof RunRevisionMount
     */
    readOnly?: boolean;
}

/**
 * Check if a given object implements the RunRevisionMount interface.
 */
export function instanceOfRunRevisionMount(value: object): value is RunRevisionMount {
    if (!('snapshotId' in value) || value['snapshotId'] === undefined) return false;
    if (!('mountPath' in value) || value['mountPath'] === undefined) return false;
    return true;
}

export function RunRevisionMountFromJSON(json: any): RunRevisionMount {
    return RunRevisionMountFromJSONTyped(json, false);
}

export function RunRevisionMountFromJSONTyped(json: any, ignoreDiscriminator: boolean): RunRevisionMount {
    if (json == null) {
        return json;
    }
    return {
        
        'snapshotId': json['snapshot_id'],
        'mountPath': json['mount_path'],
        'readOnly': json['read_only'] == null ? undefined : json['read_only'],
    };
}

export function RunRevisionMountToJSON(json: any): RunRevisionMount {
    return RunRevisionMountToJSONTyped(json, false);
}

export function RunRevisionMountToJSONTyped(value?: RunRevisionMount | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'snapshot_id': value['snapshotId'],
        'mount_path': value['mountPath'],
        'read_only': value['readOnly'],
    };
}