// src/apispec/src/models/RunUpdateRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sandbox0 API
 * Public HTTP APIs exposed by Sandbox0 regional gateways and the global gateway.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { RunScalePolicy } from './RunScalePolicy';
import {
    RunScalePolicyFromJSON,
    RunScalePolicyFromJSONTyped,
    RunScalePolicyToJSON,
    RunScalePolicyToJSONTyped,
} from './RunScalePolicy';

/**
 * 
 * @export
 * @interface RunUpdateRequest
 */
export interface RunUpdateRequest {
    /**
     * 
     * @type {string}
     * @memberof RunUpdateRequest
     */
    name?: string;
    /**
     * 
     * @type {boolean}
     * @memberof RunUpdateRequest
     */
    enabled?: boolean;
    /**
     * 
     * @type {RunScalePolicy}
     * @memberof RunUpdateRequest
     */
    scale?: RunScalePolicy;
}

/**
 * Check if a given object implements the RunUpdateRequest interface.
 */
export function instanceOfRunUpdateRequest(value: object): value is RunUpdateRequest {
    return true;
}

export function RunUpdateRequestFromJSON(json: any): RunUpdateRequest {
    return RunUpdateRequestFromJSONTyped(json, false);
}

export function RunUpdateRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): RunUpdateRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'enabled': json['enabled'] == null ? undefined : json['enabled'],
        'scale': json['scale'] == null ? undefined : RunScalePolicyFromJSON(json['scale']),
    };
}

export function RunUpdateRequestToJSON(json: any): RunUpdateRequest {
    return RunUpdateRequestToJSONTyped(json, false);
}

export function RunUpdateRequestToJSONTyped(value?: RunUpdateRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'enabled': value['enabled'],
        'scale': RunScalePolicyToJSON(value['scale']),
    };
}